requirements.txt
discord.py
requests
python-dotenv
aiohttp
google-generativeai
